requirements.txt
colorama==0.4.6
Faker==33.1.0
greenlet==3.1.1
numpy==2.1.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
